feat: bring file interface into package
I've elected to bring the file interface into this package itself to decouple from the ipfs codebase. I'm still thinking we'll track pretty closely with the IPFS definition of a file, but reduces the number of imports & provides greater islolation between the two codebases. This commit also brings a bunch of documentation fixes & updates in preparation for open sourcing.