feat(IPFS api): add options to enable IPFS API, pubsub
It's high-time we made the IPFS api re-available. This adds configuration params that, when active, create an instance of an API at the address specified by the IPFS config file at Addresses.API. While we're in here, it's worth adding support for pubsub via configuration, as some users have requested that option.

I've also formalized the "Option" interface for configuring an ipfs filestore, and added a function for configuration via a map[string]interface{}, which the qri configuration will hold.