refactor(cafs.Filestore): work with strings instead of datastore.Key
I read this: https://research.swtch.com/deps and was like "lol I already know this". After some thought, these lessons are always worth revisiting. For a long time our mental model of Qri has been one where datasets really only "exist" when they're in IPFS. That assumption is wrong. Datasets are documents, and they should be able to exist in lots of places. Because I didn't have this clarity on how important it was to think of datasets as documents regardless of their storage or representation, I made a lazy choice when designing the cafs.Filestore interface, allowing our dependency on IPFS to escape the abstraction the cafs interface was supposed to provide. The concrete result of this is our code imports "github.com/ipfs/go-datastore" when we should just be passing around strings. datastore interface details are a concern of IPFS, which the cafs interfaces must abstract away.

The cognitive results of this slip up are worse. Thinking about datasets as only being "real" when they're in IPFS holds back a whole host of benefits and workflows I've only been able to understand after lots of effort. We need to write documentation that drives home the point that datasets are documents. Doesn't matter where they're stored.

while in here I'm also deleting the gcloud package and cache interfaces because they're bad code and shouldn't be used. I really should have done this in a separate commit, but I'm not a computer.

BREAKING CHANGE:

cafs.Filestore interface methods accept and return string values instead of datastore.Key.